feat: new page for icbd5 and rename correctly previous icbd events
