+++ src/pages/events/event-router.ts
@@ -1,13 +1,17 @@
 import express from 'express'
-import  { icbdPage as icbdPageOld } from './icbd2/icbd-page'
-import  { icbdPage } from './icbd3/icbd-page'
+
+import  { icbdPage as icbdPageOld3 } from './icbd3/icbd-page'
+import  { icbdPage as icbdPageOld4 } from './icbd4/icbd-page'
+import  { icbdPage } from './icbd5/icbd-page'
 
 const eventRouter = express.Router()
 
 eventRouter.use("/", express.static("src/pages/events/assets/"))
 
-eventRouter.use(["/icbd", "/icboostday", "/icbd3"], icbdPage.render)
+eventRouter.use(["/icbd", "/icboostday", "/icbd5"], icbdPage.render)
 
-eventRouter.use(["/icbd2"], icbdPageOld.render)
+eventRouter.use(["/icbd4"], icbdPageOld4.render)
+
+eventRouter.use(["/icbd3"], icbdPageOld3.render)
 
 export { eventRouter }
