+++ src/pages/events/icbd5/icbd-labos.json
@@ -1,0 +1,8 @@
+[
+    {
+        "name": "Name PhD",
+        "role": "Role",
+        "imageURL": "/img/icbd/icbd_logo.png",
+        "websiteURL": "https://people.epfl.ch/"
+    }
+]
