+++ src/pages/events/icbd5/icbd-speakers.json
@@ -1,0 +1,8 @@
+[
+    {
+        "name": "Speaker",
+        "role": "Entreprise - Talk Name",
+        "imageURL": "/img/icbd/icbd_logo.png",
+        "websiteURL": "https://www.linkedin.com/"
+    }
+]
